fix: Fail if build fails!
The pipefail option does not catch usage from subshells like this:

```
echo <<<"$(echo foo; exit 1)"
```

Since that's how `nix develop` is called in the script, when the action
could not build an environment, it would still be treated as having run
successfully, and the job would continue.

It's still hypothetically possible for this to happen if `env` or `bash
-c 'echo $PATH'` fail.  Seems unlikely!  But maybe there's something to
be done there.